refactor(package): textwather classes into utils/textwatchers
